requirements.txt
Flask
Werkzeug
Jinja2
MarkupSafe
itsdangerous
click
blinker
python-dotenv
watchdog
pymongo